// src/vehicle/entities/vehicle.enum.ts
export enum VehicleColor {
  White = 'White',
  Black = 'Black',
  Red = 'Red',
  Blue = 'Blue',
  Silver = 'Silver',
  Gray = 'Gray',
  Green = 'Green',
  Yellow = 'Yellow',
}

export enum VehicleBrand {
  Toyota = 'Toyota',
  Ford = 'Ford',
  Honda = 'Honda',
  Chevrolet = 'Chevrolet',
  BMW = 'BMW',
  Mercedes = 'Mercedes',
  Audi = 'Audi',
  Nissan = 'Nissan',
  Volkswagen = 'Volkswagen',
  Hyundai = 'Hyundai',
  Fiat = 'Fiat',
  Yamaha = 'Yamaha',
}